@@ -11,10 +11,10 @@ license = { file = "LICENSE" }
 authors = [
   { name = "Tigrbl Team" }
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10,<3.13"
 dependencies = [
   "duckdb>=1.0.0",
-  "tigrbl>=0.0.0",
+  "tigrbl>=0.3.0.dev4",
 ]
 
 [project.urls]
