@@ -3,10 +3,11 @@ from __future__ import annotations
 from typing import Any, Mapping, Tuple
 from urllib.parse import quote_plus
 
-from sqlalchemy import create_engine, text
+from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
 SessionFactory = sessionmaker
+
 
 def _pg_url(mapping: Mapping[str, Any] | None, dsn: str | None) -> str:
     if dsn:
@@ -16,8 +17,8 @@ def _pg_url(mapping: Mapping[str, Any] | None, dsn: str | None) -> str:
     host = str(mapping.get("host") or "localhost")
     port = int(mapping.get("port") or 5432)
     user = quote_plus(str(mapping.get("user") or ""))
-    pwd  = quote_plus(str(mapping.get("pwd") or ""))
-    db   = quote_plus(str(mapping.get("db") or ""))
+    pwd = quote_plus(str(mapping.get("pwd") or ""))
+    db = quote_plus(str(mapping.get("db") or ""))
     params = []
     app = mapping.get("application_name")
     if app:
@@ -28,11 +29,21 @@ def _pg_url(mapping: Mapping[str, Any] | None, dsn: str | None) -> str:
     q = f"?{'&'.join(params)}" if params else ""
     return f"postgresql+psycopg://{user}:{pwd}@{host}:{port}/{db}{q}"
 
-def build_postgres_wal_engine(*, mapping: Mapping[str, Any] | None = None, spec: Any | None = None,
-                              dsn: str | None = None, **_) -> Tuple[Any, SessionFactory]:
+
+def build_postgres_wal_engine(
+    *,
+    mapping: Mapping[str, Any] | None = None,
+    spec: Any | None = None,
+    dsn: str | None = None,
+    **_,
+) -> Tuple[Any, SessionFactory]:
     url = _pg_url(mapping, dsn)
-    pool_size    = int((mapping or {}).get("pool_size")    or getattr(spec, "pool_size", 10) or 10)
-    max_overflow = int((mapping or {}).get("max_overflow") or getattr(spec, "max", 20)       or 20)
+    pool_size = int(
+        (mapping or {}).get("pool_size") or getattr(spec, "pool_size", 10) or 10
+    )
+    max_overflow = int(
+        (mapping or {}).get("max_overflow") or getattr(spec, "max", 20) or 20
+    )
 
     eng = create_engine(
         url,
@@ -44,6 +55,7 @@ def build_postgres_wal_engine(*, mapping: Mapping[str, Any] | None = None, spec:
 
     # Optional: set session characteristics on first connection (no server restarts)
     from sqlalchemy import event
+
     @event.listens_for(eng, "connect")
     def _configure_postgres(dbapi_conn, conn_record):
         # These are safe per-connection tweaks; WAL is a server config, not set here.
@@ -53,6 +65,7 @@ def build_postgres_wal_engine(*, mapping: Mapping[str, Any] | None = None, spec:
             cur.execute("SET synchronous_commit = 'on'")
 
     return eng, sessionmaker(bind=eng, expire_on_commit=False)
+
 
 def postgres_wal_capabilities() -> dict[str, Any]:
     return {
