@@ -1,8 +1,14 @@
 from __future__ import annotations
 from dataclasses import dataclass
-from typing import Any, Optional
+from typing import Any
 
-from tigrbl_acme_ca.key_mgmt.providers import FileKeyProvider, KmsKeyProvider, Pkcs11KeyProvider, KeyProvider
+from tigrbl_acme_ca.key_mgmt.providers import (
+    FileKeyProvider,
+    KmsKeyProvider,
+    Pkcs11KeyProvider,
+    KeyProvider,
+)
+
 
 @dataclass
 class CaKeyLoader:
@@ -13,11 +19,18 @@ class CaKeyLoader:
         km_cfg = (config or {}).get("acme.ca", {})
         source = (km_cfg.get("key_source") or "file").lower()
         if source == "file":
-            prov = FileKeyProvider(path=km_cfg.get("key_path", "ca.key.pem"), password=km_cfg.get("key_password"))
+            prov = FileKeyProvider(
+                path=km_cfg.get("key_path", "ca.key.pem"),
+                password=km_cfg.get("key_password"),
+            )
         elif source == "kms":
-            prov = KmsKeyProvider(key_id=km_cfg.get("kms_key_id"), region=km_cfg.get("kms_region"))
+            prov = KmsKeyProvider(
+                key_id=km_cfg.get("kms_key_id"), region=km_cfg.get("kms_region")
+            )
         elif source == "pkcs11":
-            prov = Pkcs11KeyProvider(slot=int(km_cfg.get("slot", 0)), label=km_cfg.get("label"))
+            prov = Pkcs11KeyProvider(
+                slot=int(km_cfg.get("slot", 0)), label=km_cfg.get("label")
+            )
         else:
             raise ValueError(f"Unsupported key_source: {source}")
         return cls(provider=prov)
