@@ -8,7 +8,7 @@ try:
     from cryptography import x509
     from cryptography.hazmat.primitives import hashes, serialization
     from cryptography.hazmat.primitives.asymmetric import rsa, ec
-    from cryptography.x509.oid import NameOID, ExtensionOID
+    from cryptography.x509.oid import ExtensionOID
 except Exception:  # pragma: no cover
     x509 = None
     hashes = None
@@ -16,12 +16,14 @@ except Exception:  # pragma: no cover
     rsa = None
     ec = None
 
+
 @dataclass
 class IssueResult:
     pem: str
     serial_hex: str
     not_before: datetime
     not_after: datetime
+
 
 @dataclass
 class CertService:
@@ -35,12 +37,25 @@ class CertService:
     def from_config(cls, config: dict) -> "CertService":
         return cls()
 
-    async def issue_certificate(self, *, csr_der: bytes, sans: List[str], issuer_key: Any,
-                                not_before: datetime, not_after: datetime) -> IssueResult:
-        import secrets, base64
+    async def issue_certificate(
+        self,
+        *,
+        csr_der: bytes,
+        sans: List[str],
+        issuer_key: Any,
+        not_before: datetime,
+        not_after: datetime,
+    ) -> IssueResult:
+        import secrets
+        import base64
+
         if x509 is None or serialization is None:
             # minimal placeholder (non-functional cert)
-            pem = "-----BEGIN CERTIFICATE-----\n" + base64.b64encode(csr_der).decode("ascii") + "\n-----END CERTIFICATE-----\n"
+            pem = (
+                "-----BEGIN CERTIFICATE-----\n"
+                + base64.b64encode(csr_der).decode("ascii")
+                + "\n-----END CERTIFICATE-----\n"
+            )
             return IssueResult(
                 pem=pem,
                 serial_hex=secrets.token_hex(16),
@@ -55,7 +70,9 @@ class CertService:
         builder = (
             x509.CertificateBuilder()
             .subject_name(csr.subject)
-            .issuer_name(csr.subject)  # In dev, default to self-issued unless issuer cert provided elsewhere
+            .issuer_name(
+                csr.subject
+            )  # In dev, default to self-issued unless issuer cert provided elsewhere
             .public_key(csr.public_key())
             .serial_number(serial)
             .not_valid_before(not_before)
@@ -64,12 +81,15 @@ class CertService:
 
         # Copy SAN from CSR if present; otherwise use provided sans
         try:
-            ext = csr.extensions.get_extension_for_oid(ExtensionOID.SUBJECT_ALTERNATIVE_NAME)
+            ext = csr.extensions.get_extension_for_oid(
+                ExtensionOID.SUBJECT_ALTERNATIVE_NAME
+            )
             builder = builder.add_extension(ext.value, critical=False)
         except Exception:
             if sans:
                 builder = builder.add_extension(
-                    x509.SubjectAlternativeName([x509.DNSName(n) for n in sans]), critical=False
+                    x509.SubjectAlternativeName([x509.DNSName(n) for n in sans]),
+                    critical=False,
                 )
 
         # Sign with issuer key
@@ -78,7 +98,9 @@ class CertService:
             cert = builder.sign(private_key=issuer_key, algorithm=algorithm)
         except Exception:
             # If issuer_key is an adapter without .sign(), fall back to placeholder
-            return await self.issue_certificate_placeholder(csr_der=csr_der, not_before=not_before, not_after=not_after)
+            return await self.issue_certificate_placeholder(
+                csr_der=csr_der, not_before=not_before, not_after=not_after
+            )
 
         pem = cert.public_bytes(serialization.Encoding.PEM).decode("ascii")
         return IssueResult(
@@ -88,9 +110,17 @@ class CertService:
             not_after=not_after,
         )
 
-    async def issue_certificate_placeholder(self, *, csr_der: bytes, not_before: datetime, not_after: datetime) -> IssueResult:
-        import secrets, base64
-        pem = "-----BEGIN CERTIFICATE-----\n" + base64.b64encode(csr_der).decode("ascii") + "\n-----END CERTIFICATE-----\n"
+    async def issue_certificate_placeholder(
+        self, *, csr_der: bytes, not_before: datetime, not_after: datetime
+    ) -> IssueResult:
+        import secrets
+        import base64
+
+        pem = (
+            "-----BEGIN CERTIFICATE-----\n"
+            + base64.b64encode(csr_der).decode("ascii")
+            + "\n-----END CERTIFICATE-----\n"
+        )
         return IssueResult(
             pem=pem,
             serial_hex=secrets.token_hex(16),
