@@ -2,7 +2,6 @@ from __future__ import annotations
 from fastapi import HTTPException
 from tigrbl_acme_ca.tables.authorizations import Authorization
 
-from fastapi import HTTPException
 
 def _h(ctx, name: str):
     handlers = ctx.get("handlers") or {}
@@ -11,11 +10,14 @@ def _h(ctx, name: str):
         raise HTTPException(status_code=500, detail=f"handler_unavailable:{name}")
     return fn
 
+
 def _id(obj):
     return obj.get("id") if isinstance(obj, dict) else getattr(obj, "id", None)
 
+
 def _field(obj, name: str):
     return obj.get(name) if isinstance(obj, dict) else getattr(obj, name, None)
+
 
 async def get_authorization(ctx, authorization_id: str):
     read_by_id = _h(ctx, "table.read.by_id")
