@@ -4,7 +4,6 @@ from typing import Any, Dict
 from fastapi import HTTPException
 from tigrbl_acme_ca.tables.accounts import Account
 
-from fastapi import HTTPException
 
 def _h(ctx, name: str):
     handlers = ctx.get("handlers") or {}
@@ -13,11 +12,14 @@ def _h(ctx, name: str):
         raise HTTPException(status_code=500, detail=f"handler_unavailable:{name}")
     return fn
 
+
 def _id(obj):
     return obj.get("id") if isinstance(obj, dict) else getattr(obj, "id", None)
 
+
 def _field(obj, name: str):
     return obj.get(name) if isinstance(obj, dict) else getattr(obj, name, None)
+
 
 async def key_change(ctx) -> Dict[str, Any]:
     p = ctx.get("payload") or {}
@@ -34,6 +36,8 @@ async def key_change(ctx) -> Dict[str, Any]:
         raise HTTPException(status_code=404, detail="account_not_found")
 
     if _field(acct, "key_thumbprint") != new_key_thumbprint:
-        await update(table=Account, id=account_id, values={"key_thumbprint": new_key_thumbprint})
+        await update(
+            table=Account, id=account_id, values={"key_thumbprint": new_key_thumbprint}
+        )
 
     return {"account_id": str(account_id), "key_thumbprint": new_key_thumbprint}
