@@ -1,5 +1,6 @@
 from __future__ import annotations
-import base64, hashlib
+import base64
+import hashlib
 from datetime import datetime, timezone
 
 from fastapi import HTTPException
@@ -9,7 +10,6 @@ from tigrbl_acme_ca.tables.authorizations import Authorization
 from tigrbl_acme_ca.tables.orders import Order
 from tigrbl_acme_ca.tables.accounts import Account
 
-from fastapi import HTTPException
 
 def _h(ctx, name: str):
     handlers = ctx.get("handlers") or {}
@@ -18,11 +18,14 @@ def _h(ctx, name: str):
         raise HTTPException(status_code=500, detail=f"handler_unavailable:{name}")
     return fn
 
+
 def _id(obj):
     return obj.get("id") if isinstance(obj, dict) else getattr(obj, "id", None)
 
+
 def _field(obj, name: str):
     return obj.get(name) if isinstance(obj, dict) else getattr(obj, name, None)
+
 
 async def post_challenge(ctx, challenge_id: str | None = None):
     read_by_id = _h(ctx, "table.read.by_id")
@@ -54,7 +57,11 @@ async def post_challenge(ctx, challenge_id: str | None = None):
             raise HTTPException(status_code=500, detail="dns_resolver_unavailable")
         name = f"_acme-challenge.{domain.rstrip('.')}"
         answers = await resolver.txt_lookup(name)
-        expected = base64.urlsafe_b64encode(hashlib.sha256(key_auth.encode("utf-8")).digest()).rstrip(b"=").decode("ascii")
+        expected = (
+            base64.urlsafe_b64encode(hashlib.sha256(key_auth.encode("utf-8")).digest())
+            .rstrip(b"=")
+            .decode("ascii")
+        )
         candidates = {a.strip().strip('"') for a in (answers or [])}
         ok = expected in candidates
     elif _field(ch, "type") == "http-01":
@@ -63,12 +70,16 @@ async def post_challenge(ctx, challenge_id: str | None = None):
             raise HTTPException(status_code=500, detail="http_client_unavailable")
         url = f"http://{domain}/.well-known/acme-challenge/{token}"
         status, body = await http.get(url)
-        ok = (status == 200 and (body or "").strip() == key_auth)
+        ok = status == 200 and (body or "").strip() == key_auth
     else:
         raise HTTPException(status_code=400, detail="unsupported_challenge_type")
 
     if ok:
-        await update(table=Challenge, id=challenge_id, values={"status": "valid", "validated_at": datetime.now(timezone.utc)})
+        await update(
+            table=Challenge,
+            id=challenge_id,
+            values={"status": "valid", "validated_at": datetime.now(timezone.utc)},
+        )
         await update(table=Authorization, id=_id(authz), values={"status": "valid"})
         authzs = await read_list(table=Authorization, where={"order_id": _id(order)})
         if authzs and all((_field(a, "status") == "valid") for a in authzs):
