@@ -10,6 +10,7 @@ from tigrbl_acme_ca.services.audit.redact import redact_payload
 
 from fastapi import HTTPException
 
+
 def _h(ctx, name: str):
     handlers = ctx.get("handlers") or {}
     fn = handlers.get(name)
@@ -17,11 +18,14 @@ def _h(ctx, name: str):
         raise HTTPException(status_code=500, detail=f"handler_unavailable:{name}")
     return fn
 
+
 def _id(obj):
     return obj.get("id") if isinstance(obj, dict) else getattr(obj, "id", None)
 
+
 def _field(obj, name: str):
     return obj.get(name) if isinstance(obj, dict) else getattr(obj, name, None)
+
 
 def _actor_from_ctx(ctx) -> Optional[str]:
     return ctx.get("actor_account_id")
@@ -33,14 +37,17 @@ async def _audit_account_created(cls, ctx):
     if not acct:
         return
     create = _h(ctx, "table.create")
-    await create(table=AuditEvent, values={
-        "event_type": "account.created",
-        "actor_account_id": _actor_from_ctx(ctx),
-        "object_type": "account",
-        "object_id": _id(acct),
-        "payload": redact_payload(ctx.get("payload")),
-        "at": datetime.now(timezone.utc),
-    })
+    await create(
+        table=AuditEvent,
+        values={
+            "event_type": "account.created",
+            "actor_account_id": _actor_from_ctx(ctx),
+            "object_type": "account",
+            "object_id": _id(acct),
+            "payload": redact_payload(ctx.get("payload")),
+            "at": datetime.now(timezone.utc),
+        },
+    )
 
 
 @hook_ctx(ops=("new_order",), phase="POST_COMMIT")
@@ -49,14 +56,17 @@ async def _audit_order_created(cls, ctx):
     if not order:
         return
     create = _h(ctx, "table.create")
-    await create(table=AuditEvent, values={
-        "event_type": "order.created",
-        "actor_account_id": _actor_from_ctx(ctx),
-        "object_type": "order",
-        "object_id": _id(order),
-        "payload": redact_payload(ctx.get("payload")),
-        "at": datetime.now(timezone.utc),
-    })
+    await create(
+        table=AuditEvent,
+        values={
+            "event_type": "order.created",
+            "actor_account_id": _actor_from_ctx(ctx),
+            "object_type": "order",
+            "object_id": _id(order),
+            "payload": redact_payload(ctx.get("payload")),
+            "at": datetime.now(timezone.utc),
+        },
+    )
 
 
 @hook_ctx(ops=("finalize",), phase="POST_COMMIT")
@@ -65,14 +75,17 @@ async def _audit_certificate_issued(cls, ctx):
     if not order:
         return
     create = _h(ctx, "table.create")
-    await create(table=AuditEvent, values={
-        "event_type": "certificate.issued",
-        "actor_account_id": _actor_from_ctx(ctx),
-        "object_type": "order",
-        "object_id": _id(order),
-        "payload": {"certificate_id": _field(order, "certificate_id")},
-        "at": datetime.now(timezone.utc),
-    })
+    await create(
+        table=AuditEvent,
+        values={
+            "event_type": "certificate.issued",
+            "actor_account_id": _actor_from_ctx(ctx),
+            "object_type": "order",
+            "object_id": _id(order),
+            "payload": {"certificate_id": _field(order, "certificate_id")},
+            "at": datetime.now(timezone.utc),
+        },
+    )
 
 
 @hook_ctx(ops=("revoke_cert",), phase="POST_COMMIT")
@@ -81,11 +94,14 @@ async def _audit_certificate_revoked(cls, ctx):
     if not rev:
         return
     create = _h(ctx, "table.create")
-    await create(table=AuditEvent, values={
-        "event_type": "certificate.revoked",
-        "actor_account_id": _actor_from_ctx(ctx),
-        "object_type": "certificate",
-        "object_id": _field(rev, "certificate_id"),
-        "payload": redact_payload(ctx.get("payload")),
-        "at": datetime.now(timezone.utc),
-    })
+    await create(
+        table=AuditEvent,
+        values={
+            "event_type": "certificate.revoked",
+            "actor_account_id": _actor_from_ctx(ctx),
+            "object_type": "certificate",
+            "object_id": _field(rev, "certificate_id"),
+            "payload": redact_payload(ctx.get("payload")),
+            "at": datetime.now(timezone.utc),
+        },
+    )
