@@ -1,16 +1,16 @@
 from __future__ import annotations
 
 import datetime as dt
-from uuid import UUID
 
 from tigrbl.orm.tables import Base
 from tigrbl.orm.mixins import GUIDPk, Timestamped
 from tigrbl.specs import acol, IO, F, S, ColumnSpec
-from tigrbl.column.storage_spec import ForeignKeySpec
 from tigrbl.types import (
-    PgUUID, UUID as UUIDType, String, Boolean, TZDateTime, JSONB, Text, Integer,
-    SAEnum, LargeBinary, Index
+    String,
+    Boolean,
+    TZDateTime,
 )
+
 
 class ExternalAccountBinding(Base, GUIDPk, Timestamped):
     __tablename__ = "acme_external_account_bindings"
@@ -20,27 +20,29 @@ class ExternalAccountBinding(Base, GUIDPk, Timestamped):
         spec=ColumnSpec(
             storage=S(type_=String(120), unique=True, index=True, nullable=False),
             field=F(py_type=str),
-            io=IO(in_verbs=("create",), out_verbs=("read","list"), sortable=True),
+            io=IO(in_verbs=("create",), out_verbs=("read", "list"), sortable=True),
         )
     )
     hmac_key_b64 = acol(
         spec=ColumnSpec(
             storage=S(type_=String(512), nullable=False),
             field=F(py_type=str),
-            io=IO(in_verbs=("create",), out_verbs=("read","list")),
+            io=IO(in_verbs=("create",), out_verbs=("read", "list")),
         )
     )
     enabled = acol(
         spec=ColumnSpec(
             storage=S(type_=Boolean, nullable=False, default=True),
             field=F(py_type=bool),
-            io=IO(in_verbs=("create","update"), out_verbs=("read","list")),
+            io=IO(in_verbs=("create", "update"), out_verbs=("read", "list")),
         )
     )
     expires_at = acol(
         spec=ColumnSpec(
             storage=S(type_=TZDateTime, nullable=True),
-            field=F(py_type=dt.datetime, allow_null_in=("create","update")),
-            io=IO(in_verbs=("create","update"), out_verbs=("read","list"), sortable=True),
+            field=F(py_type=dt.datetime, allow_null_in=("create", "update")),
+            io=IO(
+                in_verbs=("create", "update"), out_verbs=("read", "list"), sortable=True
+            ),
         )
     )
