@@ -1,14 +1,15 @@
 from __future__ import annotations
 
 from contextlib import contextmanager
-from typing import Optional
 
 # Try to use OpenTelemetry if available; otherwise no-op.
 try:
     from opentelemetry import trace
     from opentelemetry.sdk.trace import TracerProvider
     from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
-    from opentelemetry.trace.propagation.tracecontext import TraceContextTextMapPropagator
+    from opentelemetry.trace.propagation.tracecontext import (
+        TraceContextTextMapPropagator,
+    )
 except Exception:  # pragma: no cover
     trace = None
     TracerProvider = None
@@ -17,6 +18,7 @@ except Exception:  # pragma: no cover
     TraceContextTextMapPropagator = None
 
 _tracer = None
+
 
 def setup_tracing(service_name: str = "tigrbl-acme-ca") -> None:
     global _tracer
@@ -27,6 +29,7 @@ def setup_tracing(service_name: str = "tigrbl-acme-ca") -> None:
     provider.add_span_processor(BatchSpanProcessor(ConsoleSpanExporter()))
     trace.set_tracer_provider(provider)
     _tracer = trace.get_tracer(service_name)
+
 
 @contextmanager
 def start_span(name: str):
@@ -39,6 +42,7 @@ def start_span(name: str):
         finally:
             pass
 
+
 def inject_trace_headers(headers: dict) -> None:
     if TraceContextTextMapPropagator is None:
         return
@@ -49,6 +53,7 @@ def inject_trace_headers(headers: dict) -> None:
         headers.update(carrier)
     except Exception:
         pass
+
 
 def record_exception(span, exc: Exception) -> None:
     if span is None:
