@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-import base64, hashlib
+import base64
+import hashlib
 from datetime import datetime, timezone
 
 from fastapi import HTTPException
@@ -10,7 +11,6 @@ from tigrbl_acme_ca.tables.challenges import Challenge
 from tigrbl_acme_ca.tables.authorizations import Authorization
 from tigrbl_acme_ca.tables.orders import Order
 
-from fastapi import HTTPException
 
 def _h(ctx, name: str):
     handlers = ctx.get("handlers") or {}
@@ -19,14 +19,18 @@ def _h(ctx, name: str):
         raise HTTPException(status_code=500, detail=f"handler_unavailable:{name}")
     return fn
 
+
 def _id(obj):
     return obj.get("id") if isinstance(obj, dict) else getattr(obj, "id", None)
+
 
 def _field(obj, name: str):
     return obj.get(name) if isinstance(obj, dict) else getattr(obj, name, None)
 
+
 def _b64url(data: bytes) -> str:
     return base64.urlsafe_b64encode(data).rstrip(b"=").decode("ascii")
+
 
 async def validate_challenge_task(ctx, *, challenge_id: str) -> dict:
     read_by_id = _h(ctx, "table.read.by_id")
@@ -41,6 +45,7 @@ async def validate_challenge_task(ctx, *, challenge_id: str) -> dict:
     order = await read_by_id(table=Order, id=_field(authz, "order_id"))
 
     from tigrbl_acme_ca.tables.accounts import Account
+
     acct = await read_by_id(table=Account, id=_field(order, "account_id"))
 
     domain = _field(authz, "identifier")
@@ -64,14 +69,20 @@ async def validate_challenge_task(ctx, *, challenge_id: str) -> dict:
             raise HTTPException(status_code=500, detail="http_client_unavailable")
         url = f"http://{domain}/.well-known/acme-challenge/{token}"
         status, body = await http.get(url)
-        ok = (status == 200 and (body or "").strip() == key_auth)
+        ok = status == 200 and (body or "").strip() == key_auth
     else:
         raise HTTPException(status_code=400, detail="unsupported_challenge_type")
 
     if ok:
-        await update(table=Challenge, id=challenge_id, values={"status": "valid", "validated_at": datetime.now(timezone.utc)})
+        await update(
+            table=Challenge,
+            id=challenge_id,
+            values={"status": "valid", "validated_at": datetime.now(timezone.utc)},
+        )
         await update(table=Authorization, id=_id(authz), values={"status": "valid"})
-        all_authzs = await read_list(table=Authorization, where={"order_id": _id(order)})
+        all_authzs = await read_list(
+            table=Authorization, where={"order_id": _id(order)}
+        )
         if all_authzs and all((_field(a, "status") == "valid") for a in all_authzs):
             await update(table=Order, id=_id(order), values={"status": "ready"})
     else:
@@ -80,10 +91,12 @@ async def validate_challenge_task(ctx, *, challenge_id: str) -> dict:
 
     return {"challenge_id": str(challenge_id), "status": "valid" if ok else "invalid"}
 
+
 async def sweep_expired_nonces_task(ctx) -> int:
     delete = _h(ctx, "table.delete.where")
     now = datetime.now(timezone.utc)
     return await delete(table=Nonce, where={"expires_at__lte": now})
+
 
 async def sweep_expired_authzs_orders_task(ctx) -> dict:
     read_list = _h(ctx, "table.read.list")
